automatic_status_update/requirements.txt
beautifulsoup4==4.14.3
python-dotenv==1.2.2
requests==2.32.5